+++ src/main/java/com/iexec/blockchain/command/generic/CommandStorage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,9 +17,7 @@
 package com.iexec.blockchain.command.generic;
 
 import com.iexec.blockchain.api.CommandStatus;
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;
 
 import java.time.Instant;
@@ -69,7 +67,6 @@ public abstract class CommandStorage<C extends Command<A>, A extends CommandArgs
     public boolean updateToProcessing(String chainObjectId) {
         Optional<C> localCommand = commandRepository
                 .findByChainObjectId(chainObjectId)
-                .filter(command -> command.getStatus() != null)
                 .filter(command -> command.getStatus() == CommandStatus.RECEIVED);
         if (localCommand.isEmpty()) {
             return false;
@@ -90,32 +87,25 @@ public abstract class CommandStorage<C extends Command<A>, A extends CommandArgs
      * @param receipt       blockchain receipt
      */
     public void updateToFinal(String chainObjectId,
-                              @NonNull TransactionReceipt receipt) {
-        Optional<C> localCommand = commandRepository
+                              TransactionReceipt receipt) {
+        final C command = commandRepository
                 .findByChainObjectId(chainObjectId)
-                .filter(command -> command.getStatus() != null)
-                .filter(command -> command.getStatus() == CommandStatus.PROCESSING);
-        if (localCommand.isEmpty()) {
+                .filter(cmd -> cmd.getStatus() == CommandStatus.PROCESSING)
+                .orElse(null);
+        if (command == null) {
+            log.error("No entry was found in database, could not update to final state");
             return;
         }
-        C command = localCommand.get();
 
-        CommandStatus status;
-        if (StringUtils.isNotEmpty(receipt.getStatus())
-                && receipt.getStatus().equals("0x1")) {
-            status = CommandStatus.SUCCESS;
-            log.info("Success command with transaction receipt " +
-                            "[chainObjectId:{}, command:{}, receipt:{}]",
-                    chainObjectId,
-                    command.getClass().getSimpleName(),
-                    receipt);
+        if (receipt != null && receipt.isStatusOK()) {
+            command.setStatus(CommandStatus.SUCCESS);
+            log.info("Success command with transaction receipt [chainObjectId:{}, command:{}, receipt:{}]",
+                    chainObjectId, command.getClass().getSimpleName(), receipt);
         } else {
-            status = CommandStatus.FAILURE;
-            log.info("Failure after transaction sent [chainObjectId:{}, " +
-                            "command:{}, receipt:{}]", chainObjectId,
-                    command.getClass().getSimpleName(), receipt);
+            command.setStatus(CommandStatus.FAILURE);
+            log.info("Failure after transaction sent [chainObjectId:{}, command:{}, receipt:{}]",
+                    chainObjectId, command.getClass().getSimpleName(), receipt);
         }
-        command.setStatus(status);
         command.setTransactionReceipt(receipt);
         command.setFinalDate(Instant.now());
         commandRepository.save(command);